encoder/backend/requirements.txt
fastapi~=0.79.0
starlette~=0.19.1
pydantic~=1.9.1
uvicorn
mysql-connector-python